requirements.txt
altair==4.2.2
asttokens==2.2.1
async-generator==1.10
attrs==21.4.0
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
blinker==1.6.2
Brotli==1.0.9
cachetools==5.0.0
category-encoders==2.4.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
chart-studio==1.1.0
click==8.0.4
colorama==0.3.9
constantly==15.1.0
cryptography==37.0.1
cssselect==1.1.0
cycler==0.11.0
Cython==0.29.28
dash==2.3.0
dash-auth==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datareader==0.0.9
decorator==5.1.1
distlib==0.3.4
distro==1.8.0
emoji==1.6.1
entrypoints==0.4
ephem==4.1.4
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.6.0
Flask==2.0.3
Flask-Compress==1.11
Flask-SeaSurf==0.3.1
fonttools==4.31.2
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
h11==0.13.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.3
importlib-metadata==6.6.0
incremental==21.3.0
ipython==8.13.2
ipyvizzu==0.15.0
itemadapter==0.5.0
itemloaders==1.0.4
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
jmespath==1.0.0
joblib==1.1.0
jsonschema==4.4.0
jupyter-core==4.9.2
kiwisolver==1.4.1
korean-lunar-calendar==0.3.1
lxml==4.8.0
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
nbformat==5.2.0
nltk==3.7
numpy==1.22.3
opencv-python==4.5.5.64
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pandas==1.4.1
pandas-datareader==0.10.0
parsel==1.6.0
parso==0.8.3
patsy==0.5.2
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.5.1
plotly==5.6.0
plotly-express==0.4.1
prompt-toolkit==3.0.38
Protego==0.2.1
protobuf==3.19.4
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.1b0
PyDispatcher==2.0.5
Pygments==2.15.1
PyMeeus==0.5.12
pymongo==4.1.1
Pympler==1.0.1
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pywin32==303
queuelib==1.6.2
regex==2022.6.2
requests==2.27.1
requests-file==1.5.1
retrying==1.3.3
rich==13.3.5
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.0
Scrapy==2.6.1
seaborn==0.12.2
selenium==4.1.3
service-identity==21.1.0
setuptools-git==1.2
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.6.2
statsmodels==0.13.2
streamlit==1.22.0
tabula-py==2.6.0
target-encoding==1.1.0
tenacity==8.0.1
threadpoolctl==3.1.0
tldextract==3.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.64.0
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
typing_extensions==4.2.0
tzdata==2023.3
tzlocal==5.0.1
ua-parser==0.10.0
uritemplate==4.1.1
urllib3==1.26.9
validators==0.20.0
virtualenv==20.14.0
w3lib==1.22.0
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.0.3
wordnet==0.0.1b2
wsproto==1.1.0
xgboost==1.6.1
zipp==3.15.0
zope.interface==5.4.0